rename
・getRegisteredCourse -> getCourseById
・deleteRegisteredCourse -> deleteCourse
